Different time ranges warning added in custom comparing form
